backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
openai-whisper
groq
ffmpeg-python
httpx
gtts
langdetect